Refatory userService

+++ src/user/user.service.ts
@@ -1,29 +1,55 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
-import { hash } from 'bcryptjs';
+import {
+  BadRequestException,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
+import { compare, hash } from 'bcryptjs';
 import { User } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
 
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
+import { CreateUserDto } from './dto/createUser.dto';
+import { UpdateUserDto } from './dto/updateUser.dto';
 import { MessagesHelper } from './helpers/messages.helper';
+import { RoleService } from 'src/role/role.service';
+import { ChangePasswordDto } from './dto/changePassword.dto';
 
 @Injectable()
 export class UserService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly roleService: RoleService,
+  ) {}
 
-  async create(createUserDto: CreateUserDto): Promise<User> {
+  async isAuthorizationUser(
+    urerRequestId: string,
+    roleNewUserLevel: number,
+  ): Promise<boolean> {
+    const userRequest = await this.prisma.user.findUnique({
+      where: {
+        id: urerRequestId,
+      },
+      include: { role: true },
+    });
+    if (userRequest.role.level === 0) {
+      return true;
+    } else if (userRequest.role.level === 1 && roleNewUserLevel <= 1) {
+      return false;
+    }
+    return true;
+  }
+
+  async create(userId: string, createUserDto: CreateUserDto): Promise<User> {
+    const existsRole = await this.roleService.findById(createUserDto.roleId);
+
+    if (!this.isAuthorizationUser(userId, existsRole.level)) {
+      throw new UnauthorizedException(MessagesHelper.UNAUTHORIZED_CREATE);
+    }
+
     const existsUser = await this.prisma.user.findUnique({
       where: { email: createUserDto.email },
     });
     if (existsUser) {
       throw new BadRequestException(MessagesHelper.USER_EMAIL_EXISTS);
-    }
-
-    const existsRole = await this.prisma.role.findUnique({
-      where: { id: createUserDto.roleId },
-    });
-    if (!existsRole) {
-      throw new BadRequestException(MessagesHelper.USER_ROLEID_NOT_EXIST);
     }
 
     const passwordHashed = await hash(createUserDto.password, 6);
@@ -48,18 +74,17 @@ export class UserService {
     return usersWithoutPassword;
   }
 
-  async findById(id: string): Promise<User> {
+  async findById(userId: string): Promise<User> {
     const existsUser = await this.prisma.user.findUnique({
-      where: { id },
+      where: { id: userId },
     });
     if (!existsUser) {
       throw new BadRequestException(MessagesHelper.USER_ID_NOT_FOUND);
     }
-    const user = {
+    return {
       ...existsUser,
       password: undefined,
     };
-    return user;
   }
 
   async findByEmail(email: string): Promise<User> {
@@ -69,38 +94,89 @@ export class UserService {
     if (!existsUser) {
       throw new BadRequestException(MessagesHelper.USER_EMAIL_NOT_FOUND);
     }
-    const user = {
+
+    return {
       ...existsUser,
       password: undefined,
     };
-    return user;
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto) {
+  async update(userId: string, id: string, updateUserDto: UpdateUserDto) {
+    if (updateUserDto.email || updateUserDto.password) {
+      throw new BadRequestException(
+        MessagesHelper.USER_UPDATE_NOT_PASSWORD_OR_EMAIL,
+      );
+    }
     const existsUser = await this.prisma.user.findUnique({
       where: { id },
     });
     if (!existsUser) {
       throw new BadRequestException(MessagesHelper.USER_ID_NOT_FOUND);
     }
+
+    if (updateUserDto.roleId) {
+      const existsRole = await this.roleService.findById(updateUserDto.roleId);
+      if (!this.isAuthorizationUser(userId, existsRole.level)) {
+        throw new UnauthorizedException(MessagesHelper.UNAUTHORIZED_UPDATE);
+      }
+    }
+
     const updatedUser = await this.prisma.user.update({
       where: { id },
       data: updateUserDto,
     });
 
-    return updatedUser;
+    return {
+      ...updatedUser,
+      password: undefined,
+    };
   }
 
-  async remove(id: string): Promise<string> {
+  async remove(userId: string, id: string): Promise<string> {
     const existsUser = await this.prisma.user.findUnique({
       where: { id },
     });
     if (!existsUser) {
       throw new BadRequestException(MessagesHelper.USER_ID_NOT_FOUND);
     }
+
+    const userToDelete = await this.prisma.user.findUnique({
+      where: {
+        id,
+      },
+      include: { role: true },
+    });
+
+    if (!this.isAuthorizationUser(userId, userToDelete.role.level)) {
+      throw new UnauthorizedException(MessagesHelper.UNAUTHORIZED_DELETE);
+    }
+
     await this.prisma.user.delete({
       where: { id },
     });
     return `Esta ação removeu a #${id} user`;
+  }
+
+  async changePassword(
+    userId: string,
+    { password, newPassword }: ChangePasswordDto,
+  ): Promise<string> {
+    const user = await this.prisma.user.findUnique({
+      where: { id: userId },
+    });
+    const MatchesPassword = await compare(password, user.password);
+    if (!MatchesPassword) {
+      throw new BadRequestException(
+        MessagesHelper.USER_UPDATE_PASSWORD_NOT_MATCH,
+      );
+    }
+    const newPasswordHashed = await hash(newPassword, 6);
+
+    await this.prisma.user.update({
+      where: { id: userId },
+      data: { password: newPasswordHashed },
+    });
+
+    return MessagesHelper.USER_PASSWORD_UPDATE_SUCCESS;
   }
 }
